Added Chart, Removed Warnings, Cleaned the code

+++ frontend/src/components/QAndASection/DeleteQuestion.js
@@ -3,6 +3,7 @@ import { useRouteMatch, useHistory } from 'react-router-dom'
 import { selectQuestion, deleteQuestion } from '../../api/apiFBQA'
 import { QDeleteForm } from './Forms/QDeleteForm'
 import { useSelector } from "react-redux";
+import Swal from "sweetalert2";
 // import deleteQuestionBg from './ImagesD/deleteQuestionBg.png';
 
 export const DeleteQuestion = () =>{
@@ -24,8 +25,20 @@ export const DeleteQuestion = () =>{
     },[])
  
     const onSubmit = async(data) =>{
-        await deleteQuestion(data, match.params.id)
-        history.push(`/member/QandA/${member_email}`);
+        
+        Swal.fire({
+            title: 'Confirm the delete process!',
+            showCancelButton: true,
+            confirmButtonText: 'Delete',
+          }).then((result) => {
+            if (result.isConfirmed) {
+                deleteQuestion(data, match.params.id)
+                    Swal.fire('The Question has been deleted!', '', 'success')
+                    history.push(`/member/QandA/${member_email}`);
+            } else {
+              Swal.fire('The delete process has been canceled!', '', 'info')
+            }
+          })
     }
 
     return question ?(
